move user operations file
